Add menubar with links, Allow switch between different preset duration
