+++ app/controllers/dashboards_controller.rb
@@ -17,6 +17,6 @@ class DashboardsController < ApplicationController
     team_id = params[:team_id].present? ? params[:team_id] : current_user.team_id
     @team = Team.includes(:sub_teams).find team_id
 
-    @duration = params[:duration].present? ? params[:duration] : "last_seven_days"
+    @duration = params[:duration].present? ? params[:duration] : ::Dashboard::VALID_DURATIONS.first[0]
   end
 end
