+++ app/views/shared/components/_dropdown_menu.html.erb
@@ -1,0 +1,20 @@
+<details class="flex items-center">
+  <summary class="flex cursor-pointer items-center justify-between select_language p-2">
+    <h2 class="flex flex-col justify-start text-base">
+      <span class="text-primary"><%= selected_label %></span>
+    </h2>
+    <span class="icon icon-down-arrow-green icon-small"></span>
+    <span class="icon-up-arrow-green hidden icon-small"></span>
+  </summary>
+  <ul class="absolute bg-white border border-line-colour-light mt-2 rounded-lg box-shadow-small  w-[178px]">
+    <% menu.each do |item| %>
+      <li>
+        <% if item.respond_to?(:link) %>
+          <%= link_to item.label, item.link, class: 'block p-4 text-letter-color-light text-sm border-b border-line-colour bg-slate-grey-light' %>
+        <% else %>
+          item.label
+        <% end %>
+      </li>
+    <% end %>
+  </ul>
+</details>
